requirements.txt
openai
pydantic
fastapi
uvicorn
faiss-cpu
numpy
python-docx


